requirements.txt
aiohttp==3.8.5
beautifulsoup4==4.12.2
feedparser==6.0.10
langchain==0.0.267
openai==0.27.8
faiss-cpu==1.7.4
motor==3.2.0
praw==7.7.1
python-dotenv==1.0.0
slack-sdk==3.21.3
tweepy==4.14.0
jinja2==3.1.2
